Feature1/Audio Classification/requirements.txt
openai-whisper
gradio
sentence-transformers
numpy
